Complete front-end part of visualisation server
The plotting part is now taking an array of objects containing
sensor value data in a specific format, that matches the format
that will be used in the sqlite back-end. It gets grouped into
separate arrays for each available sensor, and plotted together
in a chart, where the x-axis is time.

The idea is that it will only be necessary to create a database
query for a set of data from the database, and the result  can
then directly be passed to the visualisation logic.